Refactor the project structure by moving JavaScript modules into directories and updating the necessary imports/paths.
* **Move JavaScript files**: Move `accordion.js`, `background.js`, `content.js`, `dataHandlers.js`, `fetchContent.js`, `markdown.js`, `markdownUtils.js`, `popup.js`, `settings.js`, and `uiComponents.js` to `scripts/components/`.
* **Update imports in HTML files**: Update script paths in `markdown.html`, `settings-html` and `popup.html` to reflect the new directory structure.
