--- fetchContent.js
@@ -1,23 +1,0 @@
-/*
- * This file is part of Markdown Collector.
- *
- * Markdown Collector is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Markdown Collector is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Markdown Collector.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-
-browser.runtime.onMessage.addListener((request, sender, sendResponse) => {
-    if (request.command === "getPageContent") {
-      sendResponse({ html: document.documentElement.outerHTML });
-    }
-  });
