+++ settings.html
@@ -339,6 +339,6 @@
     </div>
   </div>
   
-  <script src="settings.js"></script>
+  <script src="scripts/components/settings.js"></script>
 </body>
 </html>
